+++ src/base/model.js
@@ -4,9 +4,7 @@ import _, { assign, identity, mapKeys, mapValues, clone } from 'lodash';
 import inherits from 'inherits';
 
 import Events from './events';
-
-const PIVOT_PREFIX = '_pivot_';
-const DEFAULT_TIMESTAMP_KEYS = ['created_at', 'updated_at'];
+import {PIVOT_PREFIX, DEFAULT_TIMESTAMP_KEYS} from '../constants';
 
 // List of attributes attached directly from the `options` passed to the constructor.
 const modelProps = ['tableName', 'hasTimestamps'];
@@ -113,16 +111,20 @@ ModelBase.prototype.initialize = function() {};
  *
  * This tells the model which attribute to expect as the unique identifier
  * for each database row (typically an auto-incrementing primary key named
- * `"id"`). Note that if you are using {@link Model#parse parse} and {@link
+ * `'id'`). Note that if you are using {@link Model#parse parse} and {@link
  * Model#format format} (to have your model's attributes in `camelCase`,
  * but your database's columns in `snake_case`, for example) this refers to
- * the name returned by parse (`myId`), not the database column (`my_id`).
+ * the name returned by parse (`myId`), not the actual database column
+ * (`my_id`).
  *
- * If the table you're working with does not have an Primary-Key in the form
- * of a single column - you'll have to override it with a getter that returns
- * null. (overriding with undefined does not cascade the default behavior of
- * the value `'id'`.
- * Such a getter in ES6 would look like `get idAttribute() { return null }`
+ * You can also get the parsed id attribute value by using the model's
+ * {@link Model#parsedIdAttribute parsedIdAttribute} method.
+ *
+ * If the table you're working with does not have a Primary-Key in the form
+ * of a single column you'll have to override it with a getter that returns
+ * `null`. Overriding with `undefined` does not cascade the default behavior of
+ * the value `'id'`. Such a getter in ES6 would look like
+ * `get idAttribute() { return null }`.
  */
 ModelBase.prototype.idAttribute = 'id';
 
@@ -147,6 +149,31 @@ ModelBase.prototype.hasTimestamps = false;
 
 /**
  * @method
+ * @private
+ * @description
+ *
+ * Converts the timestamp keys to actual Date objects. This will not run if the
+ * model doesn't have {@link Model#hasTimestamps hasTimestamps} set to either
+ * `true` or an array of key names.
+ * This method is run internally when reading data from the database to ensure
+ * data consistency between the several database implementations.
+ * It returns the model instance that called it, so it allows chaining of other
+ * model methods.
+ *
+ * @returns {Model} The model that called this.
+ */
+ModelBase.prototype.formatTimestamps = function formatTimestamps() {
+  if (!this.hasTimestamps) return this;
+
+  this.getTimestampKeys().forEach((key) => {
+    this.set(key, new Date(this.get(key)));
+  });
+
+  return this;
+};
+
+/**
+ * @method
  * @description  Get the current value of an attribute from the model.
  * @example      note.get("title");
  *
@@ -156,6 +183,35 @@ ModelBase.prototype.hasTimestamps = false;
 ModelBase.prototype.get = function(attr) {
   return this.attributes[attr];
 };
+
+/**
+ * @method
+ * @private
+ * @description
+ *
+ * Returns the model's {@link Model#idAttribute idAttribute} after applying the
+ * model's {@link Model#parse parse} method to it. Doesn't mutate the original
+ * value of {@link Model#idAttribute idAttribute} in any way.
+ *
+ * @example
+ *
+ * var Customer = bookshelf.Model.extend({
+ *   idAttribute: 'id',
+ *   parse: function(attrs) {
+ *     return _.mapKeys(attrs, function(value, key) {
+ *       return 'parsed_' + key;
+ *     });
+ *   }
+ * });
+ *
+ * customer.parsedIdAttribute() // 'parsed_id'
+ *
+ * @returns {mixed} Whatever value the parse method returns.
+ */
+ModelBase.prototype.parsedIdAttribute = function() {
+  var parsedAttributes = this.parse({[this.idAttribute]: null})
+  return parsedAttributes && Object.keys(parsedAttributes)[0]
+}
 
 /**
  * @method
@@ -190,7 +246,10 @@ ModelBase.prototype.set = function(key, val, options) {
   const prev    = this._previousAttributes;
 
   // Check for changes of `id`.
-  if (this.idAttribute in attrs) this.id = attrs[this.idAttribute];
+  if (this.idAttribute in attrs)
+    this.id = attrs[this.idAttribute];
+  else if (this.parsedIdAttribute() in attrs)
+    this.id = attrs[this.parsedIdAttribute()];
 
   // For each `set` attribute, update or delete the current value.
   for (const attr in attrs) {
@@ -352,16 +411,16 @@ ModelBase.prototype.has = function(attr) {
  * opposite operation of `parse`.
  *
  * @example
- *
- * // Example of a "parse" to convert snake_case to camelCase, using `underscore.string`
+ * // Example of a parser to convert snake_case to camelCase, using lodash
+ * // This is just an example. You can use the built-in case-converter plugin
+ * // to achieve the same functionality.
  * model.parse = function(attrs) {
- *   return _.reduce(attrs, function(memo, val, key) {
- *     memo[_.camelCase(key)] = val;
- *     return memo;
- *   }, {});
+ *   return _.mapKeys(attrs, function(value, key) {
+ *     return _.camelCase(key);
+ *   });
  * };
  *
- * @param {Object} response Hash of attributes to parse.
+ * @param {Object} attributes Hash of attributes to parse.
  * @returns {Object} Parsed attributes.
  */
 ModelBase.prototype.parse = identity;
@@ -492,6 +551,22 @@ ModelBase.prototype.saveMethod = function({ method = null, patch = false } = {})
 
 /**
  * @method
+ * @private
+ * @description
+ *
+ * Returns the automatic timestamp key names set on this model. Note that this
+ * will always return a value even if the model has {@link Model#hasTimestamps
+ * hasTimestamps} set to `false`. In this case and when set to `true` the
+ * return value will be the default names of `created_at` and `updated_at`.
+ *
+ * @returns {Array<string>} The two timestamp key names.
+ */
+ModelBase.prototype.getTimestampKeys = function() {
+  return Array.isArray(this.hasTimestamps) ? this.hasTimestamps : DEFAULT_TIMESTAMP_KEYS;
+}
+
+/**
+ * @method
  * @description
  * Sets the timestamp attributes on the model, if {@link Model#hasTimestamps
  * hasTimestamps} is set to `true` or an array. Check if the model {@link
@@ -517,14 +592,9 @@ ModelBase.prototype.timestamp = function(options) {
   const now          = (options || {}).date ? new Date(options.date) : new Date();
   const attributes   = {};
   const method       = this.saveMethod(options);
-  const keys = _.isArray(this.hasTimestamps)
-    ? this.hasTimestamps
-    : DEFAULT_TIMESTAMP_KEYS;
-
   const canEditUpdatedAtKey = (options || {}).editUpdatedAt!= undefined ? options.editUpdatedAt : true;
   const canEditCreatedAtKey = (options || {}).editCreatedAt!= undefined ? options.editCreatedAt : true;
-
-  const [ createdAtKey, updatedAtKey ] = keys;
+  const [ createdAtKey, updatedAtKey ] = this.getTimestampKeys();
 
   if (updatedAtKey && canEditUpdatedAtKey) {
     attributes[updatedAtKey] = now;
@@ -538,7 +608,6 @@ ModelBase.prototype.timestamp = function(options) {
 
   return attributes;
 };
-
 
 /**
  * @method