+++ test/index.js
@@ -6,27 +6,23 @@ Promise.onPossiblyUnhandledRejection(function (err) {
 });
 
 global.testPromise = Promise;
-var testQueryCache = global.testQueryCache = [];
 var oldIt = it;
 
 it = function() {
-  testQueryCache = [];
   return oldIt.apply(this, arguments);
 };
 
 // http://bluebirdjs.com/docs/api/error-management-configuration.html#global-rejection-events
 process.on("unhandledRejection", function(reason, promise) {
-    console.error(reason);
+  console.error(reason);
 });
 
 var Bookshelf = require('../bookshelf');
 var base = require('./base');
-
 global.sinon = require('sinon');
-
 var chai = global.chai = require('chai');
+var databaseConnections;
 
-// chai.use(require('chai-as-promised'));
 chai.use(require('sinon-chai'));
 chai.should();
 
@@ -35,32 +31,43 @@ global.AssertionError = chai.AssertionError;
 global.Assertion      = chai.Assertion;
 global.assert         = chai.assert;
 
-describe('Bookshelf', function () {
+after(function() {
+  return databaseConnections.forEach(function(connection) {
+    return connection.knex.destroy();
+  })
+});
 
-  it('VERSION should equal version number in package.json',
-    function () {
+describe('Bookshelf', function () {
+  it('VERSION should equal version number in package.json', function() {
     var Knex = require('knex');
     var bookshelf = Bookshelf(Knex({client: 'sqlite3', useNullAsDefault: true}));
     var p = require('../package.json');
+
     expect(p.version).to.equal(bookshelf.VERSION);
+
+    return bookshelf.knex.destroy()
   });
 
+  describe('Construction', function() {
+    it('should fail without a knex instance', function() {
+      expect(() => Bookshelf()).to.throw(/Invalid knex/);
+    });
+
+    it('should fail if passing a random object', function() {
+      expect(() => Bookshelf({config: 'something', options: ['one', 'two']})).to.throw(/Invalid knex/);
+    })
+  });
 });
 
 // Unit test all of the abstract base interfaces
 describe('Unit Tests', function () {
-
   base.Collection();
-  base.Model();
   base.Events();
-  base.Relation();
-  base.Eager();
 
   require('./unit/sql/sync')();
   require('./unit/sql/model')();
-
 });
 
 describe('Integration Tests', function () {
-  require('./integration')(Bookshelf);
+  databaseConnections = require('./integration')(Bookshelf);
 });