+++ test/integration/helpers/config.js
@@ -1,5 +1,4 @@
 module.exports = {
-
   mysql: {
     database: 'bookshelf_test',
     user: 'root',
@@ -13,14 +12,5 @@ module.exports = {
 
   sqlite3: {
     filename: ':memory:'
-  },
-
-  oracledb: {
-    user          : "travis",
-    password      : "travis",
-    connectString : "localhost/XE",
-    // https://github.com/oracle/node-oracledb/issues/525
-    stmtCacheSize : 0
   }
-
 };