+++ test/unit/sql/sync.js
@@ -1,47 +1,66 @@
-var Promise   = testPromise;
-var _         = require('lodash');
+var _ = require('lodash');
+var path = require('path');
+var basePath = process.cwd();
 
 module.exports = function() {
-
-  // This module is included into the `bookshelf` repository,
-  // and run from the root of the directory.
-  var path     = require('path');
-  var basePath = process.cwd();
-
   var Sync = require(path.resolve(basePath + '/lib/sync'));
 
   describe('Sync', function() {
-
-    var stubSync = function() {
+    var stubModel = function(idAttribute) {
       var qd = [];
-      var stubQuery = function() {
-        qd.push(_.toArray(arguments));
-        return this;
-      };
+
       return {
+        idAttribute: idAttribute || 'id',
+        id: 'pk',
+        attributes: {
+          idAttribute: 'pk'
+        },
         tableName: 'testtable',
+        format: _.identity,
         isNew: function() {
           return true
         },
         queryData: qd,
-        query: function() {
-          return {
-            where: stubQuery,
-            limit: stubQuery
-          };
+        operation: null,
+        query: function() { return this._query },
+        _query: {
+          _statements: qd,
+          where: function(where) { qd.push({grouping: 'where', where}) },
+          limit: function(limit) { qd.push({grouping: 'limit', limit}) },
         },
-        resetQuery: function () {
+        resetQuery: function() {
           return this;
+        },
+        getWhereParts: function() {
+          return qd
+          .filter(function(item) {
+            return item.grouping == 'where'
+          })
+          .map(function(item) {
+            return item.where
+          });
         }
       };
     };
 
-    describe('prefixFields', function () {
+    it('accepts a withSchema option', function() {
+      var testSchema = 'test';
+      var setSchema = sinon.spy();
+      var mockModel = {
+        query: function() {
+          return {withSchema: setSchema}
+        },
+        resetQuery: function() {}
+      }
 
-      it('should prefix all keys of the passed in object with the tablename', function () {
+      new Sync(mockModel, {withSchema: testSchema});
 
-        var sync = new Sync(stubSync());
+      setSchema.should.have.been.calledWith(testSchema);
+    })
 
+    describe('prefixFields', function() {
+      it('should prefix all keys of the passed in object with the tablename', function() {
+        var sync = new Sync(stubModel());
         var attributes = {
           'some': 'column',
           'another': 'column'
@@ -51,16 +70,14 @@ module.exports = function() {
           'testtable.some':'column',
           'testtable.another':'column'
         });
-
       });
 
-      it('should run after format', function() {
-
+      it('should run after format for select', function() {
         var attributes = {
           'Some': 'column',
           'Another': 'column'
         };
-        var sync = new Sync(_.extend(stubSync(), {
+        var sync = new Sync(_.extend(stubModel(), {
           format: function(attrs) {
             var data = {};
             for (var key in attrs) {
@@ -69,20 +86,95 @@ module.exports = function() {
             return data;
           }
         }));
+
         sync.select = function() {
-          expect(this.syncing.queryData[0]).to.eql([{
+          expect(this.syncing.queryData[0].where).to.eql({
             "testtable.some": "column",
             "testtable.another": "column"
-          }]);
+          });
         };
-        return sync.first(attributes);
 
+        return sync.first(attributes);
       });
 
+      it('should format attributes for updates, including id attribute', function(done) {
+        var snakeCase = _.snakeCase;
+        var stubModelInstance = _.extend(stubModel('idAttribute'), {
+          format: function(attrs) {
+            var data = {};
+            for (var key in attrs) {
+              data[snakeCase(key)] = attrs[key];
+            }
+            return data;
+          }
+        });
+        var updateFields = {
+          someColumn: 'updated',
+          otherColumn: 'updated'
+        };
+
+        stubModelInstance._query.update = function(attrs) {
+          expect(stubModelInstance.getWhereParts()).to.eql([{id_attribute: 'pk'}]);
+          expect(attrs).to.eql({'some_column': 'updated', 'other_column': 'updated'});
+          done()
+        }
+
+        var sync = new Sync(stubModelInstance);
+        sync.update(updateFields);
+      });
+
+      it('should format id attribute for deletes', function(done) {
+        var snakeCase = _.snakeCase;
+        var stubModelInstance = _.extend(stubModel('idAttribute'), {
+          idAttribute: 'idAttribute',
+          format: function(attrs) {
+            var data = {};
+            for (var key in attrs) {
+              data[snakeCase(key)] = attrs[key];
+            }
+            return data;
+          }
+        })
+
+        stubModelInstance._query.del = function() {
+          expect(stubModelInstance.getWhereParts()).to.eql([{id_attribute: 'pk'}]);
+          done()
+        }
+
+        var sync = new Sync(stubModelInstance);
+        sync.del()
+      });
     });
 
+    describe('update', function() {
+      it('doesn\'t try to update the primary key if it hasn\'t changed', function() {
+        var sync = new Sync(stubModel());
+        _.extend(sync.query, {
+          update: function(attrs) {
+            expect(attrs).to.not.have.property('id');
+          },
+          where: function() {
+            this._statements = [{grouping: 'where'}];
+          }
+        });
 
+        return sync.update({id: 'pk', name: 'something'});
+      });
+
+      it('will update the primary key if it has changed', function() {
+        var sync = new Sync(stubModel());
+        _.extend(sync.query, {
+          update: function(attrs) {
+            expect(attrs).to.have.property('id');
+            expect(attrs.id).to.equal('updated');
+          },
+          where: function() {
+            this._statements = [{grouping: 'where'}];
+          }
+        });
+
+        return sync.update({id: 'updated', name: 'something'});
+      })
+    })
   });
-
-
 };